+++ src/Resources/contao/dca/tl_form_field.php
@@ -1,0 +1,11 @@
+<?php
+
+$dca = &$GLOBALS['TL_DCA']['tl_form_field'];
+
+/**
+ * Palettes
+ */
+
+$dca['palettes']['multifileupload'] = '{type_legend},type,name,label;{fconfig_legend},mandatory,extensions,maxlength;{store_legend:hide},uploadFolder,useHomeDir;{expert_legend:hide},class,accesskey,tabindex,fSize;{template_legend:hide},customTpl;{invisible_legend:hide},invisible';
+
+