+++ src/Resources/contao/languages/de/tl_form_field.php
@@ -1,0 +1,5 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['FFL'];
+
+$lang['multifileupload'] = ['Datei-Upload (Drag & Drop)', 'Ein Eingabefeld zur Übertragung mehrerer lokaler Dateien mit Drag & Drop Unterstützung auf den Server.'];