+++ src/Resources/contao/languages/en/tl_form_field.php
@@ -1,0 +1,5 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['FFL'];
+
+$lang['multifileupload'] = ['File-Upload (Drag & Drop)', 'An input field for transferring several local files with drag & drop support to the server.'];